+++ src/index.tsx
@@ -1,1 +1,3 @@
 export { default } from "./MarkdownPro";
+export { defaultUrlTransform, parseCodeMetadata, parseFrontmatter } from "./MarkdownPro";
+export type { CodeTheme, Frontmatter, FrontmatterValue, MarkdownProProps, TocItem, TocOptions, UrlKind, UrlTransform } from "./MarkdownPro";